requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
python-dotenv
tensorflow
scikit-learn
python-multipart